Graph - Refactored the FcgVertex to create a generic circle vertex shape provider
