+++ Ghidra/Features/Base/data/base.theme.properties
@@ -26,6 +26,8 @@ color.fg.dialog.equates.suggestion = color.palette.hint
 color.fg.consoletextpane = color.fg
 color.fg.consoletextpane.error = color.fg.error
 
+color.bg.graph.vertex.function = MediumAquamarine
+
 color.fg.infopanel.version = color.fg
 
 color.bg.interpreterconsole = color.bg
