+++ Ghidra/Features/Base/src/main/java/ghidra/base/graph/VertexExpansionListener.java
@@ -1,0 +1,38 @@
+/* ###
+ * IP: GHIDRA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ghidra.base.graph;
+
+import ghidra.graph.viewer.VisualVertex;
+
+/**
+ * A listener to know when a vertex has been told to expand
+ */
+public interface VertexExpansionListener {
+
+	/**
+	* Show or hide those vertices that are on incoming edges to v
+	* 
+	* @param v the vertex
+	*/
+	public void toggleIncomingVertices(VisualVertex v);
+
+	/**
+	 * Show or hide those vertices that are on outgoing edges to v
+	 * 
+	 * @param v the vertex
+	 */
+	public void toggleOutgoingVertices(VisualVertex v);
+}
