+++ Ghidra/Features/GraphFunctionCalls/data/functioncallgraph.theme.properties
@@ -1,9 +1,8 @@
 [Defaults]
 
-color.bg.plugin.fcg.vertex.default = MediumAquamarine // TODO
 color.bg.plugin.fcg.vertex.toobig = color.palette.lightgray
 
-color.bg.plugin.fcg.edge.primary.direct = darkseagreen // TODO
+color.bg.plugin.fcg.edge.primary.direct = darkseagreen 
 color.bg.plugin.fcg.edge.primary.direct.selected = color.palette.lightgreen
 color.bg.plugin.fcg.edge.primary.indirect = color.palette.lavender 
 color.bg.plugin.fcg.edge.primary.indirect.selected = color.palette.silver
